perf: shrink EvalValue/LValue (float_type double; LValue field reorder)
The script float type changes from `long double` to `double`, and LValue's
`is_const` bool moves after `container_idx`. Together this shrinks the core
value objects:

  EvalValue: 48 -> 40 bytes   (long double forced 16-byte alignment, padding
                               EvalValue to 48; double is 8-aligned -> 40)
  LValue:    80 -> 56 bytes   (the smaller EvalValue, plus is_const packing
                               into the tail instead of wasting 7 bytes between
                               val and the container pointer)

LValue is the array element type, so the 30% drop cuts memory traffic for every
array op and every value copy. Measured speedups vs the prior build: array_append
1.31x, reverse 1.26x, sum 1.24x, foreach 1.22x/1.20x, concat 1.17x, subscript
1.13x, float_arith 1.13x, and math_builtins 2.27x (double libm is far faster
than the long double variants). Overall my/py geomean is now ~1.04x. Floats now
match Python's exactly (both 64-bit IEEE double); the only divergence left is
str() formatting (C %f fixed notation vs Python's shortest round-trip), which
the bench harness normalizes.

Tradeoff (accepted): double has ~15-16 significant digits vs long double's
~18-19, and an int beyond 2^53 mixed into a float loses exactness (long double's
64-bit mantissa held all 64-bit ints). Pure-int arithmetic is unchanged
(int_type stays intptr_t).

Mechanics: float parsing stold->stod (parser, float() builtin); the math
builtins switch from the `l`-suffixed libm functions to the double std::
overloads (selected by a function-pointer cast); fmodl/truncl/fabsl/roundl/powl
-> the double versions; str()'s snprintf format "%.*Lf" -> "%.*f" (this was also
a latent vararg bug: a double passed to a %Lf slot printed nan). size_type is
already uint32_t, so off/len needed no change. README / CLAUDE.md / bench/README
updated; the float-width divergence note is removed. 578 tests green, no
warnings, all 51 benchmark outputs match Python.

+++ src/types/float.cpp.h
@@ -80,7 +80,7 @@ void TypeFloat::mod(EvalValue &a, const EvalValue &b)
     if (std::fpclassify(rhs) == FP_ZERO)
         throw DivisionByZeroEx();
 
-    a = fmodl(a.get<float_type>(), rhs);
+    a = std::fmod(a.get<float_type>(), rhs);
 }
 
 void TypeFloat::lt(EvalValue &a, const EvalValue &b)
@@ -163,7 +163,7 @@ size_t TypeFloat::hash(const EvalValue &a)
      * representable in floating point), so [int_min, -int_min) is exactly
      * the set of integral floats that fit in int_type.
      */
-    if (std::isfinite(v) && v == truncl(v) && v >= int_min && v < -int_min)
+    if (std::isfinite(v) && v == std::trunc(v) && v >= int_min && v < -int_min)
         return std::hash<int_type>()(static_cast<int_type>(v));
 
     return std::hash<float_type>()(v);
